requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
